refactor: Certification Query, Command 리팩토링
