+++ src/main/java/coffeemeet/server/certification/implement/CertificationQuery.java
@@ -3,7 +3,8 @@ package coffeemeet.server.certification.implement;
 import static coffeemeet.server.certification.exception.CertificationErrorCode.CERTIFICATION_NOT_FOUND;
 
 import coffeemeet.server.certification.domain.Certification;
-import coffeemeet.server.certification.infrastructure.CertificationRepository;
+import coffeemeet.server.certification.domain.CompanyEmail;
+import coffeemeet.server.certification.domain.repository.CertificationRepository;
 import coffeemeet.server.common.execption.InvalidInputException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -31,6 +32,10 @@ public class CertificationQuery {
         .orElseThrow(() -> new InvalidInputException(CERTIFICATION_NOT_FOUND,
             String.format(CERTIFICATION_NOT_FOUND_MESSAGE, userId))
         ).getCompanyName();
+  }
+
+  public boolean isExistedCompanyEmail(CompanyEmail companyEmail) {
+    return certificationRepository.existsByCompanyEmail(companyEmail);
   }
 
   public Long getUserIdByCertificationId(Long certificationId) {
